R Markdown Presentation & Plotly
Instructions

Create a web page presentation using R Markdown that features a plot
created with Plotly. Host your webpage on either GitHub Pages, RPubs,
or NeoCities. Your webpage must contain the date that you created the
document, and it must contain a plot created with Plotly. We would love
to see you show off your creativity!
Review criterialess
The rubric contains the following two questions:

Does the web page feature a date and is this date less than two months
before the date that you're grading this assignment?
Is the web page a presentation and does it feature an interactive plot
that appears to have been created with Plotly?